fastapi_web/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
sqlmodel
mysqlclient
python-dotenv
pydantic
pydantic-settings